perf: 优化控制 ACL Action Choices 的选项
