+++ apps/acls/models/__init__.py
@@ -2,4 +2,3 @@ from .command_acl import *
 from .connect_method import *
 from .login_acl import *
 from .login_asset_acl import *
-from .base import ActionChoices
