+++ apps/acls/serializers/connect_method.py
@@ -1,6 +1,7 @@
 from orgs.mixins.serializers import BulkOrgResourceModelSerializer
 from .base import BaseUserAssetAccountACLSerializer as BaseSerializer
 from ..models import ConnectMethodACL
+from ..const import ActionChoices
 
 __all__ = ["ConnectMethodACLSerializer"]
 
@@ -12,12 +13,6 @@ class ConnectMethodACLSerializer(BaseSerializer, BulkOrgResourceModelSerializer)
             i for i in BaseSerializer.Meta.fields + ['connect_methods']
             if i not in ['assets', 'accounts']
         ]
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        field_action = self.fields.get('action')
-        if not field_action:
-            return
-        # 仅支持拒绝
-        for k in ['review', 'accept']:
-            field_action._choices.pop(k, None)
+        action_choices_exclude = BaseSerializer.Meta.action_choices_exclude + [
+            ActionChoices.review, ActionChoices.accept
+        ]
