+++ apps/acls/serializers/login_acl.py
@@ -2,7 +2,7 @@ from django.utils.translation import ugettext as _
 
 from common.serializers import MethodSerializer
 from orgs.mixins.serializers import BulkOrgResourceModelSerializer
-from .base import BaserUserACLSerializer
+from .base import BaseUserACLSerializer
 from .rules import RuleSerializer
 from ..models import LoginACL
 
@@ -11,12 +11,12 @@ __all__ = ["LoginACLSerializer"]
 common_help_text = _("With * indicating a match all. ")
 
 
-class LoginACLSerializer(BaserUserACLSerializer, BulkOrgResourceModelSerializer):
+class LoginACLSerializer(BaseUserACLSerializer, BulkOrgResourceModelSerializer):
     rules = MethodSerializer(label=_('Rule'))
 
-    class Meta(BaserUserACLSerializer.Meta):
+    class Meta(BaseUserACLSerializer.Meta):
         model = LoginACL
-        fields = BaserUserACLSerializer.Meta.fields + ['rules', ]
+        fields = BaseUserACLSerializer.Meta.fields + ['rules', ]
 
     def get_rules_serializer(self):
         return RuleSerializer()
